Corte1/Teoria/Proyecto Potenciometro/requirements.txt
flask
pandas
matplotlib
pyserial